Added IPv6 support and auto use if needed
Added support for IPv6. In transparent mode IPv4 may not be available thus this payload will detect if we should scan all the IPv6 addresses we can find. The version of NMap that comes with Packet Squirrel seems to not support scripts which is shame. I hope future versions will.