Togglable VPN Payload
Based on the default VPN payload; this can now create a VPN-connection to an OpenVPN-server, or if the button is pressed, send traffic from the clients through said tunnel. This way no editing of the payload is required to switch modes. On boot the Squirrel will wait for a button press for 10 seconds, if it is pressed, the VPN will launch in client mode, if it is not pressed in the interval it will launch in remote-access mode instead.